requirements.txt
pandas==2.2.3
pyarrow==15.0.2
numpy==1.26.4

PyMySQL==1.1.1

apache-airflow-providers-mysql==5.7.1

python-dotenv==1.0.1
requests==2.31.0
tenacity==8.2.3
python-dateutil==2.9.0.post0
google-cloud-storage==2.17.0
google-cloud-bigquery==3.25.0
